requirements.txt
alembic==1.15.1
asgiref==3.8.1
asttokens==3.0.0
b4==0.12.4
beautifulsoup4==4.12.3
bin2png==2.0
blinker==1.7.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
contourpy==1.2.0
coverage==7.9.2
crispy-bootstrap5==2025.4
cryptography==45.0.5
cryptography318==1.0.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.4
django-crispy-forms==2.4
djangorestframework==3.16.0
dkimpy==1.1.5
dnspython==2.6.1
executing==2.2.0
Flask==3.0.2
flask-cors==6.0.0
fonttools==4.47.2
fpdf==1.7.2
git-filter-repo==2.38.0
gunicorn==23.0.0
idna==3.10
ipython==8.32.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
jpg-png==0.0.1
kiwisolver==1.4.5
Mako==1.3.9
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
myon-png==1.3
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.3.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
patatt==0.6.3
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.2.2
plotly==5.19.0
pngpadder==0.3
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py2png==0.1.1
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.17.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.1
pyperclip==1.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.5
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
tenacity==8.2.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.0.6
whitenoise==6.9.0
